pyproject.toml
[project]
name = "umara-image-search"
version = "0.1.0"
description = "AI-powered semantic image search for e-commerce media archives"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Gemini API
    "google-genai>=1.67.0",
    # Vector store — FAISS is dependency-free, fast, and production-grade
    # (ChromaDB was first choice but its onnxruntime dependency dropped macOS x86_64)
    "faiss-cpu>=1.8.0",
    # API server
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30.0",
    # Image processing
    "pillow>=12.1.1",
    # Config
    "python-dotenv>=1.0.0",
    # Retry logic
    "tenacity>=8.2.0",
    # Nice CLI output during backfill
    "rich>=13.0.0",
]

[project.scripts]
backfill = "backfill:main"
serve = "uvicorn api.app:app --reload"

[tool.uv]
dev-dependencies = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
